Move request FSA dispatch to just after bailout call
This allows for the request actions to update things like state based de-duplication before getting the other fields.

Use Case:
Making dispatch a RSAA on componentDidMount and componentDidUpdate on a react component.  
Use the Bailout to check if the request is current.
Use the Request FSA to update the state in Redux that Bailout is reading.

This works perfectly, until you perform some action in endpoint, body, or headers, that it's self also updates the state in Redux, triggering componentDidUpdate in our component, and firing a second RSAA who check's it's bailout before the state changes, resulting in a second network call.

By moving the Request FSA to immediately after the bailout, we can update our state before anything else happens.